# openapi.yaml
openapi: 3.0.3
info:
  title: MediaHub API
  description: |
    Contrato mínimo gerado a partir de `server/server.mjs`.
    Autenticação: header `Authorization: Bearer <JWT>` quando indicado.
    Alguns endpoints de busca aceitam token opcional (chaves do provedor por usuário).
  version: "1.0.0"
servers:
  - url: http://localhost:8081
    description: API local (padrão `PORT` no .env)
  - url: https://api.example.com
    description: Substituir pela URL real em produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/search/query:
    get:
      tags: [search]
      summary: Busca multi, filme ou série
      description: Requer autenticação. Usa chaves do usuário e/ou do sistema.
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [multi, movie, tv]
          description: Padrão multi
        - in: query
          name: query
          required: true
          schema:
            type: string
            maxLength: 120
        - in: query
          name: language
          schema:
            type: string
            default: pt-BR
        - in: query
          name: year
          schema:
            type: string
            pattern: '^\d{4}$'
      responses:
        "200":
          description: Payload no formato do provedor (ex. TMDb-like)
        "400":
          description: Consulta inválida
        "401":
          description: Não autenticado
        "429":
          description: Rate limit
  /api/search/videos:
    get:
      tags: [search]
      summary: Vídeos (trailers) do título
      description: Requer autenticação.
      parameters:
        - in: query
          name: mediaType
          schema:
            type: string
            enum: [movie, tv]
        - in: query
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: language
          schema:
            type: string
      responses:
        "200":
          description: Lista de vídeos
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
  /api/search/details:
    get:
      tags: [search]
      summary: Detalhes completos do título
      parameters:
        - in: query
          name: mediaType
          schema:
            type: string
            enum: [movie, tv]
        - in: query
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: language
          schema:
            type: string
      responses:
        "200":
          description: Detalhes do provedor
  /api/search/trending:
    get:
      tags: [search]
      summary: Trending
      parameters:
        - in: query
          name: mediaType
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
      responses:
        "200":
          description: Lista trending
  /api/search/image:
    get:
      tags: [search]
      summary: Proxy de imagem do provedor
      description: Requer autenticação (Bearer ou cookie auth_token).
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
        - in: query
          name: size
          schema:
            type: string
        - in: query
          name: download
          schema:
            type: string
        - in: query
          name: filename
          schema:
            type: string
      responses:
        "200":
          description: Binário imagem
        "401":
          description: Não autenticado
...